Extract common logic for Ruby Sequel benchmarks (#2520)
This should make it easier to keep all the benchmarks in the suite
up-to-date and in-sync! And easier to compare and contrast the
frameworks.

[ci fw-only Ruby/sinatra-sequel Ruby/rack-sequel Ruby/roda-sequel]
