+++ frameworks/Ruby/rack-sequel/config.ru
@@ -1,3 +1,4 @@
+require_relative 'boot'
 require_relative 'hello_world'
 use Rack::ContentLength
 use Rack::Chunked
