+++ frameworks/Ruby/roda-sequel/config.ru
@@ -1,2 +1,3 @@
+require_relative 'boot'
 require_relative 'hello_world'
 run HelloWorld.freeze.app
