+++ frameworks/Ruby/roda-sequel/source_code
@@ -1,1 +1,3 @@
+./roda-sequel/config.ru
+./roda-sequel/boot.rb
 ./roda-sequel/hello_world.rb
